# openapi.yaml
openapi: 3.0.3
info:
  title: Evenly API
  version: 0.1.0
  description: |
    HTTP contract expected by the frontend service adapter. Money values are
    integer minor units (kobo for NGN). The frontend sends and receives JSON
    except for receipt uploads, which use multipart/form-data.
servers:
  - url: /api
    description: Configure the frontend base URL with VITE_API_BASE_URL.
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register and sign in
      security: []
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/RegisterInput' } } } }
      responses:
        '200': { description: Created session, content: { application/json: { schema: { $ref: '#/components/schemas/Session' } } } }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Sign in
      security: []
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LoginInput' } } } }
      responses:
        '200': { description: Signed-in session, content: { application/json: { schema: { $ref: '#/components/schemas/Session' } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
  /expenses:
    post:
      tags: [Expenses]
      summary: Create an expense
      security: [{ sessionCookie: [] }]
      description: Requires membership in the referenced group.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ExpenseInput' } } } }
      responses:
        '201': { description: Created expense, content: { application/json: { schema: { $ref: '#/components/schemas/Expense' } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /expenses/{expenseId}:
    parameters: [{ $ref: '#/components/parameters/ExpenseId' }]
    get:
      tags: [Expenses]
      summary: Get an expense
      security: [{ sessionCookie: [] }]
      description: Requires membership in the expense's group.
      responses:
        '200': { description: Expense, content: { application/json: { schema: { $ref: '#/components/schemas/Expense' } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Expenses]
      summary: Update an expense
      security: [{ sessionCookie: [] }]
      description: Requires membership; only the member who created it may edit it.
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ExpenseUpdateInput' } } } }
      responses:
        '200': { description: Updated expense, content: { application/json: { schema: { $ref: '#/components/schemas/Expense' } } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Expenses]
      summary: Soft-delete an expense
      security: [{ sessionCookie: [] }]
      description: Requires membership; only the member who created it may delete it.
      responses:
        '204': { description: Expense deleted }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    sessionCookie:
      type: http
      scheme: bearer
      bearerFormat: opaque token
      description: Bearer token returned by /auth/register and /auth/login.
  parameters:
...
    ExpenseId: { name: expenseId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
...
  responses:
    Unauthenticated: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Forbidden: { description: Insufficient permissions, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    Conflict: { description: Conflict, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
    ValidationError: { description: Validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  schemas:
    Id: { type: string }
    IsoDate: { type: string, format: date-time }
    CurrencyCode: { type: string, enum: [NGN] }
    Kobo: { type: integer, description: Integer amount in kobo (NGN minor units). }
    SplitMethod: { type: string, enum: [equal, custom, percentage, shares] }
...
    User:
      type: object
      required: [id, name, email, createdAt, updatedAt]
      properties: { id: { $ref: '#/components/schemas/Id' }, name: { type: string }, email: { type: string, format: email }, createdAt: { $ref: '#/components/schemas/IsoDate' }, updatedAt: { $ref: '#/components/schemas/IsoDate' } }
    Session:
      type: object
      required: [user]
      properties: { user: { oneOf: [{ $ref: '#/components/schemas/User' }, { type: 'null' }] }, access_token: { type: string, nullable: true }, token_type: { type: string, nullable: true, example: bearer } }
    RegisterInput:
      type: object
      required: [name, email, password]
      properties: { name: { type: string }, email: { type: string, format: email }, password: { type: string, format: password, minLength: 8 } }
    LoginInput:
      type: object
      required: [email, password]
      properties: { email: { type: string, format: email }, password: { type: string, format: password } }
...
    ExpensePayerInput: { type: object, required: [memberId, amount], properties: { memberId: { $ref: '#/components/schemas/Id' }, amount: { $ref: '#/components/schemas/Kobo' } } }
    ExpenseParticipantInput: { type: object, required: [memberId], properties: { memberId: { $ref: '#/components/schemas/Id' }, allocationValue: { type: number, description: Kobo for custom, percent for percentage, shares for shares; ignored for equal. } } }
    ExpenseInput:
      type: object
      required: [groupId, description, totalAmount, expenseDate, splitMethod, payers, participants]
      properties: { groupId: { $ref: '#/components/schemas/Id' }, description: { type: string }, totalAmount: { $ref: '#/components/schemas/Kobo' }, expenseDate: { type: string, format: date }, splitMethod: { $ref: '#/components/schemas/SplitMethod' }, payers: { type: array, items: { $ref: '#/components/schemas/ExpensePayerInput' } }, participants: { type: array, items: { $ref: '#/components/schemas/ExpenseParticipantInput' } }, receipt: { oneOf: [{ $ref: '#/components/schemas/UploadedFile' }, { type: 'null' }] } }
    ExpenseUpdateInput:
      type: object
      required: [description, totalAmount, expenseDate, splitMethod, payers, participants]
      properties: { description: { type: string }, totalAmount: { $ref: '#/components/schemas/Kobo' }, expenseDate: { type: string, format: date }, splitMethod: { $ref: '#/components/schemas/SplitMethod' }, payers: { type: array, items: { $ref: '#/components/schemas/ExpensePayerInput' } }, participants: { type: array, items: { $ref: '#/components/schemas/ExpenseParticipantInput' } }, receipt: { oneOf: [{ $ref: '#/components/schemas/UploadedFile' }, { type: 'null' }] } }
    UploadedFile: { type: object, required: [fileName, fileUrl], properties: { fileName: { type: string }, fileUrl: { type: string, format: uri } } }
    Receipt: { type: object, required: [id, expenseId, fileName, fileUrl, createdAt], properties: { id: { $ref: '#/components/schemas/Id' }, expenseId: { $ref: '#/components/schemas/Id' }, fileName: { type: string }, fileUrl: { type: string }, createdAt: { $ref: '#/components/schemas/IsoDate' } } }
    ExpensePayer: { $ref: '#/components/schemas/ExpensePayerInput' }
    ExpenseParticipant: { type: object, required: [memberId, allocationValue, calculatedAmount], properties: { memberId: { $ref: '#/components/schemas/Id' }, allocationValue: { type: number }, calculatedAmount: { $ref: '#/components/schemas/Kobo' } } }
    Expense:
      type: object
      required: [id, groupId, createdBy, description, totalAmount, currency, expenseDate, splitMethod, payers, participants, receipt, createdAt, updatedAt, deletedAt]
      properties: { id: { $ref: '#/components/schemas/Id' }, groupId: { $ref: '#/components/schemas/Id' }, createdBy: { $ref: '#/components/schemas/Id' }, description: { type: string }, totalAmount: { $ref: '#/components/schemas/Kobo' }, currency: { $ref: '#/components/schemas/CurrencyCode' }, expenseDate: { type: string, format: date }, splitMethod: { $ref: '#/components/schemas/SplitMethod' }, payers: { type: array, items: { $ref: '#/components/schemas/ExpensePayer' } }, participants: { type: array, items: { $ref: '#/components/schemas/ExpenseParticipant' } }, receipt: { oneOf: [{ $ref: '#/components/schemas/Receipt' }, { type: 'null' }] }, createdAt: { $ref: '#/components/schemas/IsoDate' }, updatedAt: { $ref: '#/components/schemas/IsoDate' }, deletedAt: { oneOf: [{ $ref: '#/components/schemas/IsoDate' }, { type: 'null' }] } }
...
    ApiError: { type: object, properties: { code: { type: string, enum: [validation, not_found, forbidden, unauthenticated, conflict] }, message: { type: string }, issues: { type: array, items: { $ref: '#/components/schemas/ValidationIssue' } } } }
    ValidationIssue: { type: object, required: [field, message], properties: { field: { type: string, enum: [description, totalAmount, expenseDate, payers, participants, allocation] }, message: { type: string } } }